fitness: the verdict plane crosses — six engine slices, the route, and the wire order
One route over the whole engine: /api/admin/model-fitness GET (matrix,
capabilities, health, value, detail with the live-run block and stale
sweeping, estimate, transcripts) and the POST action union — start with
its arming ladder (role checks, adversary-may-not-be-the-candidate,
concurrency 1..8), stop, clear, forget. Slices: probes, evals,
adversarial, score, value, observed, transcripts, surface.

Three port principles out of it, now crate-wide:

- Stored jsonb never rides a typed struct to the wire. Postgres keeps
  object keys in its own canonical order and TS passes the parsed blob
  through untouched, so every stored-object passthrough (the detail
  record, run rows, both views' index) rides the raw serde_json::Value;
  typed parses serve decisions only. The typed ordered map that needed
  is IndexMap, not serde_json's Map (whose methods live only on
  Map<String, Value>).
- JSON.stringify prints whole floats as integers (1, not 1.0) —
  responses pass through body::js_numberify.
- serde_json's default float parse is best-effort and can land one ULP
  off the parsed digits; a stored 0.9090909090909091 came back ...092
  and diffed inside an otherwise byte-identical 456KB body. The
  float_roundtrip feature is on for every float the API parses.

Supporting halves: the gateway registry now reads the price columns
(flat + model_prices + auto_prices, the estimate's inputs),
model_access::canonical_model_id (a stored spelling onto the offered
one, exactly one qualified pin or unchanged), capability::
forget_capabilities, and Finding's check/severity go &'static str ->
String so findings can round-trip.

The start body's zod discriminated union was probed against the
verbatim TS schema: any parse failure collapses to the union's
"Invalid input" (wrong JSON type, missing field, unknown tier, null
`only`); only bound failures surface the arm's own sentence, in field
order (model 1..200, tiers min 1 no max, adversaryModel max 200
nullish, only elements max 120 no min array max 64 optional,
concurrency int 1..8 with "expected number to be" phrasing).

Byte-diffed 27 cases against the TS oracle — 21 byte-identical, 6
exactly the app-harness boundary (app-shipped harnesses are TS-only
by rule 10; their counts, workload keys, estimate arithmetic and note
branch differ on any install with an enabled app, recorded under
divergences; the harness masks those artifacts on both sides and the
remainder must still be identical), 0 unexplained. No real run was
ever launched and clear-all never fired.

The diff also caught a TS bug, not a port one: clear 500s on TS for
every model — clearFitnessResults -> setSetting(key, null) writes SQL
NULL into app_settings.value, which is NOT NULL. Rust keeps the
working behavior (200 + counts); flagged for the TS side rather than
ported.

+++ api/src/fitness/observed.rs
@@ -240,14 +240,14 @@ async fn harness_run_groups(
       select
         harness,
         model,
-        count(*)::int                                                        as runs,
-        sum(case when schema_valid and repairs = 0 then 1 else 0 end)::int   as first_pass,
-        sum(case when schema_valid then 1 else 0 end)::int                   as held,
-        sum(case when schema_valid and repairs > 0 then 1 else 0 end)::int   as repaired,
-        sum(findings)::int                                                   as findings,
-        sum(case when widened then 1 else 0 end)::int                        as widened,
-        sum(case when error is not null then 1 else 0 end)::int              as errors,
-        coalesce(percentile_disc(0.5) within group (order by latency_ms), 0)::int as latency_p50,
+        count(*)::bigint as runs,
+        sum(case when schema_valid and repairs = 0 then 1 else 0 end)::bigint as first_pass,
+        sum(case when schema_valid then 1 else 0 end)::bigint                 as held,
+        sum(case when schema_valid and repairs > 0 then 1 else 0 end)::bigint as repaired,
+        sum(findings)::bigint                                             as findings,
+        sum(case when widened then 1 else 0 end)::bigint                    as widened,
+        sum(case when error is not null then 1 else 0 end)::bigint            as errors,
+        coalesce(percentile_disc(0.5) within group (order by latency_ms), 0)::bigint as latency_p50,
         (extract(epoch from max(created_at)) * 1000)::bigint                 as last_run_ms
       from harness_runs
       where created_at > now() - ($1 || ' days')::interval
@@ -288,7 +288,7 @@ async fn chain_step_groups(
     since_days: i64,
 ) -> Result<Vec<ChainStepGroup>, sqlx::Error> {
     let sql = r#"
-      select harness, model, chain_step as step, count(*)::int as runs
+      select harness, model, chain_step as step, count(*)::bigint as runs
       from harness_runs
       where created_at > now() - ($1 || ' days')::interval
       group by harness, model, chain_step
@@ -320,7 +320,7 @@ async fn guard_finding_groups(
     since_days: i64,
 ) -> Result<Vec<GuardFindingGroup>, sqlx::Error> {
     let sql = r#"
-      select model, check_type as "check", count(*)::int as n
+      select model, check_type as "check", count(*)::bigint as n
       from guard_findings
       where model is not null and created_at > now() - ($1 || ' days')::interval
       group by model, check_type
